+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- Hard import gate, history-consistent timestamps, first-class import/export contracts ([pm-beads-q3t0](https://github.com/unbraind/pm-beads/blob/main/.agents/pm/features/pm-beads-q3t0.toon))
+
+### Fixed
+
+- Promote pm beads validate to first-class command registration ([pm-beads-azv0](https://github.com/unbraind/pm-beads/blob/main/.agents/pm/issues/pm-beads-azv0.toon))
+
 ## 2026.06.09 - 2026-06-09
 
 ### Added